+++ packages/components/src/VirtualizedList/VirtualizedList.component.js
@@ -50,6 +50,8 @@ function VirtualizedList(props) {
 		widthsOfColumns,
 		setWidthsOfColumns,
 		headerAction,
+		headerHeight,
+		disableHeight = false,
 	} = props;
 	const columnDefinitionsWithSelection = insertSelectionConfiguration({
 		children,
@@ -91,7 +93,7 @@ function VirtualizedList(props) {
 
 	return (
 		<virtualizedListContext.Provider value={{ resizeColumn }}>
-			<AutoSizer>
+			<AutoSizer disableHeight={disableHeight}>
 				{({ height, width }) => (
 					<RendererSelector
 						ref={rendererSelectorRef}
@@ -120,6 +122,7 @@ function VirtualizedList(props) {
 						scrollToIndex={scrollToIndex}
 						scrollToAlignment={scrollToAlignment}
 						headerAction={headerAction}
+						headerHeight={headerHeight}
 					>
 						{columnDefinitions}
 					</RendererSelector>